# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "藏药小程序API",
    "description": "藏药电商小程序后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/product/list": {
      "get": {
        "tags": ["商品模块"],
        "summary": "商品列表",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {"type": "integer"}
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {"type": "string"}
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart/list": {
      "get": {
        "tags": ["购物车模块"],
        "summary": "购物车列表",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/cart/add": {
      "post": {
        "tags": ["购物车模块"],
        "summary": "添加到购物车",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {"type": "integer"},
                  "spec": {"type": "string"},
                  "quantity": {"type": "integer"}
                },
                "required": ["productId", "quantity"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/cart/update": {
      "post": {
        "tags": ["购物车模块"],
        "summary": "更新购物车",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {"type": "integer"},
                  "quantity": {"type": "integer"}
                },
                "required": ["id", "quantity"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/cart/delete": {
      "post": {
        "tags": ["购物车模块"],
        "summary": "删除购物车商品",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {"type": "integer"}
                  }
                },
                "required": ["ids"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/cart/clear": {
      "post": {
        "tags": ["购物车模块"],
        "summary": "清空购物车",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/order/list": {
      "get": {
        "tags": ["订单模块"],
        "summary": "订单列表",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          },
          {
            "name": "status",
            "in": "query",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/order/detail/{id}": {
      "get": {
        "tags": ["订单模块"],
        "summary": "订单详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/address/list": {
      "get": {
        "tags": ["地址模块"],
        "summary": "地址列表",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/address/add": {
      "post": {
        "tags": ["地址模块"],
        "summary": "添加地址",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/address/update": {
      "post": {
        "tags": ["地址模块"],
        "summary": "更新地址",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/address/delete/{id}": {
      "post": {
        "tags": ["地址模块"],
        "summary": "删除地址",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/address/setDefault/{id}": {
      "post": {
        "tags": ["地址模块"],
        "summary": "设置默认地址",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "data": {
            "type": "object",
            "properties": {
              "total": {"type": "integer"},
              "pageNum": {"type": "integer"},
              "pageSize": {"type": "integer"},
              "list": {"type": "array"}
            }
          }
        }
      },
...
      "AddressRequest": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "phone": {"type": "string"},
          "province": {"type": "string"},
          "city": {"type": "string"},
          "district": {"type": "string"},
          "detail": {"type": "string"},
          "isDefault": {"type": "boolean"}
        }
      }
    }
  }
}